+++ BionicInventory.Application/VendorPurchaseTerms/Commands/Update/UpdatedVendorPurchaseTermDto.cs
@@ -3,7 +3,7 @@
  * @Author:  Mikael Araya
  * @Contact: MikaelAraya12@gmail.com
  * @Last Modified By:  Mikael Araya
- * @Last Modified Time: Dec 26, 2018 9:37 PM
+ * @Last Modified Time: Jan 15, 2019 12:39 AM
  * @Description: Modify Here, Please 
  */
 using MediatR;
@@ -16,7 +16,7 @@ namespace BionicInventory.Application.Vendors.PurchaseTerms.Commands.Update {
         public string VendorProductId { get; set; }
         public uint? Priority { get; set; }
         public uint? Leadtime { get; set; }
-        public float? MinimumQuantity { get; set; }
+        public uint? MinimumQuantity { get; set; }
         public float UnitPrice { get; set; }
 
     }
